refactor: extract Pydantic DTO schemas into models/dto/ modules (#653)
Move all API request/response Pydantic schemas from models/schemas.py
and pipeline compilation models from models/pipeline.py into
domain-specific modules under models/dto/. This reduces import-time
coupling and allows importing DTO types without pulling in SQLModel.

New modules: dto/admin, dto/auth, dto/capture, dto/common,
dto/compilation, dto/ingest, dto/lint, dto/llm, dto/mcp, dto/query,
dto/sharing, dto/synthesis, dto/tags, dto/wiki.

models/__init__.py re-exports everything for backward compatibility.
