+++ src/wikimind/models/dto/__init__.py
@@ -1,21 +1,14 @@
-"""Domain models — SQLModel tables persisted to SQLite and Pydantic schemas for the pipeline.
+"""Dependency-light DTO schemas — re-exports for convenience.
 
-Enums define the vocabulary (source types, statuses, providers). SQLModel tables
-store sources, articles, concepts, backlinks, queries, jobs, and cost logs.
-Pydantic models carry data through the ingest -> compile -> query pipeline.
+Import DTOs from this package instead of ``wikimind.models`` when you only
+need Pydantic request/response schemas and want to avoid pulling in
+SQLModel / database dependencies.
 
-Pydantic DTOs (request/response schemas) are defined in ``wikimind.models.dto``
-and re-exported here for backward compatibility.
+Example::
 
-This package re-exports all public names so that existing imports like
-``from wikimind.models import Article`` continue to work unchanged.
+    from wikimind.models.dto import ArticleResponse, IngestURLRequest
 """
 
-from typing import Any, NamedTuple
-
-from pydantic import BaseModel
-
-# Re-export DTO request/response schemas (replaces pipeline.py + schemas.py)
 from wikimind.models.dto.admin import (
     AdminActionResult,
     AdminUserDetail,
@@ -111,6 +104,8 @@ from wikimind.models.dto.mcp import (
     MCPTokenResponse,
     MCPTokenRevokeResponse,
 )
+
+# Also re-export QueryResult from query module
 from wikimind.models.dto.query import (
     AskResponse,
     CitationArticleRef,
@@ -189,229 +184,77 @@ from wikimind.models.dto.wiki import (
     WikilinkMatch,
 )
 
-# Re-export enums
-from wikimind.models.enums import (
-    ArticleDownloadFormat,
-    CaptureKind,
-    CaptureStatus,
-    ClaimConceptRole,
-    ClusterStatus,
-    ConfidenceLevel,
-    ContradictionResolution,
-    ContradictionStatus,
-    ExportFormat,
-    IngestStatus,
-    JobStatus,
-    JobType,
-    LintFindingKind,
-    LintReportStatus,
-    LintSeverity,
-    PageType,
-    Provider,
-    RelationType,
-    SourceType,
-    TaskType,
-    WikiExportFormat,
-)
-
-# Re-export SQLModel tables from domain-specific sub-modules
-from wikimind.models.tables import (  # noqa: F401
-    Article,
-    ArticleConcept,
-    ArticleSource,
-    ArticleTag,
-    Backlink,
-    CaptureSource,
-    ClaimConcept,
-    CompilationDraft,
-    CompilationSchema,
-    CompiledClaim,
-    Concept,
-    ConceptCluster,
-    ConceptKindDef,
-    Contradiction,
-    ContradictionFinding,
-    Conversation,
-    CostLog,
-    DismissedFinding,
-    Job,
-    LintPairCache,
-    LintReport,
-    LLMTrace,
-    MCPAccessToken,
-    MigrationHistory,
-    OAuthAccessToken,
-    OAuthAuthorizationCode,
-    OrphanFinding,
-    Plan,
-    Query,
-    QueryCount,
-    ReinforcementEvent,
-    RssFeed,
-    SavedSearch,
-    ShareLink,
-    Source,
-    SourceImage,
-    StorageUsage,
-    StructuralFinding,
-    Subscription,
-    SyncLog,
-    Tag,
-    User,
-    UserApiKey,
-    UserPreference,
-    WebhookEvent,
-    _LintFindingBase,
-)
-
-# ---------------------------------------------------------------------------
-# Pydantic models that reference SQLModel tables (kept here to avoid
-# circular imports between dto/ and tables/).
-# ---------------------------------------------------------------------------
-
-
-class LintReportDetail(BaseModel):
-    """API response shape for a single report with all findings."""
-
-    report: LintReport
-    contradictions: list[ContradictionFinding]
-    orphans: list[OrphanFinding]
-    resolutions: dict[str, str] = {}  # "article_a_id|article_b_id" -> resolution
-    structurals: list[StructuralFinding] = []
-
-
-# ---------------------------------------------------------------------------
-# Typed return models that reference SQLModel tables (issue #394)
-# ---------------------------------------------------------------------------
-
-
-class FTSResponse(NamedTuple):
-    """Result of a full-text search query: paginated results and total count."""
-
-    results: list[FTSResultItem]
-    total: int
-
-
-class QAResult(NamedTuple):
-    """Result of a Q&A answer call: query row, conversation, and optional score."""
-
-    query: "Query"
-    conversation: "Conversation"
-    wiki_worthiness_score: WikiWorthinessScore | None
-
-
-class FileBackArticlePair(NamedTuple):
-    """Result of filing a conversation back to the wiki."""
-
-    article: "Article"
-    is_update: bool
-
-
-class ResolvedBacklinks(NamedTuple):
-    """Result of resolving wikilink candidates against the article table."""
-
-    resolved: list[Any]  # list[ResolvedBacklink] -- avoids circular import
-    unresolved: list[str]
-
-
-# Resolve forward references now that all DTO modules are imported.
-# SavedSearchExecuteResponse.saved_search uses forward-ref to SavedSearchResponse
-SavedSearchExecuteResponse.model_rebuild()
-
-
 __all__ = [
-    # API schemas
+    # admin
     "AdminActionResult",
     "AdminUserDetail",
     "AdminUserSummary",
-    # Pipeline models
+    # compilation
     "AnswerCompilationResult",
-    # Frontmatter
     "AnswerFrontmatter",
     "ApproveDraftRequest",
     "ApproveDraftResponse",
-    # Tables
-    "Article",
-    "ArticleConcept",
-    # Enums
-    "ArticleDownloadFormat",
+    # wiki
     "ArticleDownloadResponse",
     "ArticleEditRequest",
     "ArticleRelationshipsResponse",
     "ArticleResponse",
-    "ArticleSource",
+    # ingest
     "ArticleSourceSummary",
     "ArticleSummaryResponse",
-    "ArticleTag",
+    # tags
     "ArticleTagResponse",
+    # query
     "AskResponse",
-    "Backlink",
     "BacklinkEntry",
+    # capture
     "CaptureDiscardResponse",
     "CaptureIngestResponse",
-    "CaptureKind",
     "CaptureListResponse",
     "CaptureRequest",
     "CaptureResponse",
-    "CaptureSource",
-    "CaptureStatus",
     "CitationArticleRef",
     "CitationResponse",
-    "ClaimConcept",
-    "ClaimConceptRole",
-    "ClusterStatus",
-    "CompilationDraft",
     "CompilationDraftResponse",
     "CompilationResult",
-    "CompilationSchema",
     "CompilationSchemaResponse",
-    "CompiledClaim",
     "CompiledClaimDTO",
+    # llm
     "CompletionRequest",
     "CompletionResponse",
-    "Concept",
-    "ConceptCluster",
     "ConceptCompilationResult",
     "ConceptDetailResponse",
     "ConceptFrontmatter",
-    "ConceptKindDef",
     "ConceptResponse",
-    "ConfidenceLevel",
-    "Contradiction",
-    "ContradictionFinding",
-    "ContradictionResolution",
     "ContradictionResolutionOption",
     "ContradictionResponse",
-    "ContradictionStatus",
-    "Conversation",
     "ConversationDetail",
     "ConversationResponse",
     "ConversationSummary",
-    "CostLog",
     "CreateCompilationSchemaRequest",
     "CreateSavedSearchRequest",
+    # sharing
     "CreateShareLinkRequest",
     "CreateStubRequest",
     "CreateStubResponse",
+    # synthesis
     "CreateSynthesisRequest",
     "CreateTagRequest",
     "CrystallizeResponse",
+    # auth
     "DeleteAccountResponse",
     "DeleteConfirmation",
     "DiscardCaptureRequest",
+    # lint
     "DismissFindingResponse",
-    "DismissedFinding",
     "DocumentChunk",
     "EligibleConcept",
     "EmbeddingStats",
-    "ExportFormat",
     "ExportResponse",
-    "FTSResponse",
     "FTSResultItem",
     "FacetBucket",
     "FacetGroup",
     "FacetResponse",
-    "FileBackArticlePair",
     "FileBackArticleRef",
     "FileBackResult",
     "FileBackSelectionRequest",
@@ -422,28 +265,17 @@ __all__ = [
     "HealthReport",
     "HealthSummaryResponse",
     "IndexFrontmatter",
-    "IngestStatus",
     "IngestTextRequest",
     "IngestURLRequest",
-    "Job",
-    "JobStatus",
     "JobTriggerResponse",
-    "JobType",
     "LLMProviderStatus",
-    "LLMTrace",
     "LLMTraceListResponse",
     "LLMTraceResponse",
     "LinkedArticleSummary",
-    "LintFindingKind",
-    "LintPairCache",
-    "LintReport",
-    "LintReportDetail",
-    "LintReportStatus",
     "LintRunResponse",
-    "LintSeverity",
     "LinterContradiction",
     "LinterResult",
-    "MCPAccessToken",
+    # mcp
     "MCPTokenCreateRequest",
     "MCPTokenCreateResponse",
     "MCPTokenResponse",
@@ -453,22 +285,12 @@ __all__ = [
     "MagicLinkVerifyRequest",
     "MagicLinkVerifyResponse",
     "MetaFrontmatter",
-    "MigrationHistory",
     "NormalizedDocument",
-    "OAuthAccessToken",
-    "OAuthAuthorizationCode",
     "OAuthTokenResponse",
     "OAuthUserInfo",
     "OrphanArticle",
-    "OrphanFinding",
-    "PageType",
     "PipelineStep",
-    "Plan",
-    "Provider",
     "PublicArticleResponse",
-    "QAResult",
-    "Query",
-    "QueryCount",
     "QueryRequest",
     "QueryResponse",
     "QueryResult",
@@ -476,41 +298,28 @@ __all__ = [
     "RecentSourceEntry",
     "RecompileResponse",
     "RefreshArticleResponse",
-    "ReinforcementEvent",
     "RejectDraftResponse",
-    "RelationType",
     "RelationshipEdge",
     "ResolveContradictionBody",
     "ResolveContradictionRequest",
     "ResolveContradictionResponse",
-    "ResolvedBacklinks",
-    "RssFeed",
     "RssFeedListResponse",
     "RssFeedRequest",
     "RssFeedResponse",
     "RssFeedToggleRequest",
     "RssPollResponse",
-    "SavedSearch",
     "SavedSearchExecuteResponse",
     "SavedSearchResponse",
     "SearchResponse",
     "SearchResult",
-    "ShareLink",
     "ShareLinkResponse",
-    "Source",
     "SourceCompilationResult",
     "SourceContentResponse",
     "SourceDetailResponse",
     "SourceFrontmatter",
-    "SourceImage",
     "SourceImageEntry",
     "SourceResponse",
-    "SourceType",
-    "StorageUsage",
-    "StructuralFinding",
     "StuckSource",
-    "Subscription",
-    "SyncLog",
     "SynthesisCompilationResult",
     "SynthesisConfirmRequest",
     "SynthesisConfirmResponse",
@@ -522,24 +331,21 @@ __all__ = [
     "SynthesisResponse",
     "SynthesisSuggestion",
     "SystemStats",
-    "Tag",
     "TagArticleRequest",
     "TagResponse",
-    "TaskType",
     "TokenCreateRequest",
     "TokenCreateResponse",
     "TurnSelection",
     "TypedBacklinkSuggestion",
     "UpdateCompilationSchemaRequest",
-    "User",
-    "UserApiKey",
-    "UserPreference",
     "UserProfileResponse",
-    "WebhookEvent",
-    "WikiExportFormat",
     "WikiExportResponse",
     "WikiHealthReport",
     "WikiWorthinessScore",
     "WikilinkMatch",
     "ZombieSource",
 ]
+
+# Resolve forward references now that all DTO modules are imported.
+# SavedSearchExecuteResponse.saved_search uses forward-ref to SavedSearchResponse
+SavedSearchExecuteResponse.model_rebuild()
